refactor: 重构设置界面软件设置子页
