+++ src/renderer/components/states/maxExpand/components/setting/components/app/components/ClipboardHistorySettingsSection.tsx
